Revert "I Finished My Classes!"
This reverts commit fe124eb3d99823978c0486fe63a374d49c6be0d0.
