--- .vs/VSWorkspaceState.json
@@ -1,7 +1,0 @@
-{
-  "ExpandedNodes": [
-    ""
-  ],
-  "SelectedNode": "\\utils.py",
-  "PreviewInSolutionExplorer": false
-}
