+++ Asteriod.py
@@ -1,7 +1,16 @@
 import pyray as pr
 from random import randint
 
-class Actor:
+class Asteriod:
+
+    def __init__(self, window_width: int, window_height: int, texture: pr.Texture) -> None:
+        self.win_width = window_width
+        self.win_height = window_height
+        self.texture = texture
+        self.set_timer()
+
+    def set_timer(self) -> None:
+        self.timer = randint(0, 10)*60
 
     @staticmethod
     def create_image(character, font_size, color) -> pr.Image:
@@ -20,36 +29,10 @@ class Actor:
             [pos_x, pos_y],
             pr.WHITE
         )
-
-class Asteriod(Actor):
-
-    def __init__(self, window_width: int, window_height: int, texture: pr.Texture) -> None:
-        self.win_width = window_width
-        self.win_height = window_height
-        self.texture = texture
-        self.set_pos()
-        self.set_timer()
-
-    def draw(self) -> None:
-        self.draw_rectangle(self.texture, self.pos_x, self.pos_y)
-
-    def set_timer(self) -> None:
-        self.timer = randint(0, 10)*60
-
-    def set_pos(self) -> None:
-        self.pos_x = randint(0, self.win_width - self.texture.width)
-        self.pos_y = 0
     
     @staticmethod
     def update_actor(actor) -> None:
         actor.pos_y += 1
-    
-    @staticmethod
-    def check_collision(asteroid, player) -> bool:
-        return pr.check_collision_recs(
-                [asteroid.pos_x, asteroid.pos_y,asteroid.texture.width, asteroid.texture.height],
-                [player.pos_x, player.pos_y, player.texture.width, player.texture.height]
-            )
     
     @classmethod
     def load_asteroid(cls, amount, win_width, win_height, texture) -> None:
@@ -57,12 +40,11 @@ class Asteriod(Actor):
             cls(win_width, win_height, texture)
     
     @classmethod
-    def draw_asteroid(cls, player) -> None:
+    def draw_asteriod(cls) -> None:
         cls.timer_watch()
-        for asteroid in cls._on_screen:
-            cls.collisions_watch(asteroid, player)
+        for asteroid in cls.on_screen:
             if asteroid.pos_y >= asteroid.win_height - asteroid.texture.height:
-                cls._on_screen.remove(asteroid)
+                cls.on_screen.remove(asteroid)
                 asteroid.set_timer()
             else:
                 cls.update_actor(asteroid)
@@ -70,19 +52,13 @@ class Asteriod(Actor):
     
     @classmethod
     def timer_watch(cls) -> None:
-        for asteroid in cls._all_created:
+        for asteroid in cls.all_created:
 
-            if asteroid.timer <= 0 and asteroid not in cls._on_screen:
+            if asteroid.timer <= 0 and asteroid not in cls.on_screen:
                 asteroid.set_pos()
-                cls._on_screen.append(asteroid)
+                cls.on_screen.append(asteroid)
 
-            if asteroid.timer > 0 and asteroid in cls._on_screen:
-                cls._on_screen.remove(asteroid)
+            if asteroid.timer > 0 and asteroid in cls.on_screen:
+                cls.on_screen.remove(asteroid)
             elif asteroid.timer > 0:
                 asteroid.timer -= 1
-    
-    @classmethod
-    def collisions_watch(cls, asteroid, player) -> None:
-        if cls.check_collision(asteroid, player):
-            cls._on_screen.remove(asteroid)
-            player.points += asteroid.points
